Use encodeObject on collection elements when using encodeObject
When encodeObject is used on a collection, we should also call
encodeObject on the collection elements.  While it is possible no such
conversion is needed we do it anyways.

Update swagger example to use date type to demonstrate a case where this
conversion would come into play (coercing string elements in a map to
date)