[#59] Refactor so operations are on Cluster/Node/DC instead of Server (#60)
* Have Server.register return a 'Bound' type.
* Move operations such as prime from Server to Bound Cluster/Node/DC.
* Remove SimulacronCluster/SimulacronNode and Server generalization.   Now that we have more specific names that don't clash (like BoundCluster) this isn't needed anymore and simplifies things quite a bit.
* Move Scope to http-server and only use it there

The only thing left that Scope was tied to was the ActivityLog.  Changed
ActivityLog to be node level as it is more likely to be queried that
way.  When retrieved at DC and Cluster level it is built then.  With
that in place Scope was no longer needed, so pulled out all use of it to
Http Server.

* Update swagger doc with recent changes

Also collapse by default and add default values for ids.