[#6] Include frame in /log JSON
Adds a new 'frame' element to /log endpoint json which includes the
fully deserialized json of a native protocol frame.  This should provide
comprehensive details about messages being sent from client.

To accomplish this, added a new project, native-protocol-json that has a
sole purpose of serializing native protocol request frames to json.

Note that serializing from json to native protocol frames is not
implemented, this may be done in the future but isn't useful to us right
now.

Deprecates type, query, consistency_level, serial_consistency_level,
and client_timestamp in /log output as this is all available in the
frame.  These fields will be removed in a future release.